refactor: sprint 41 phase 3a — Wasp/Actionable/Repository extraction
Phase 3 part 1 of the models split (audit item 1.1, rows 29-31). The
biggest classes by line count leave _legacy.py and move to focused
domain files.

- libinv/models/wasp.py (new): Wasp + Wasp-only helpers.
- libinv/models/actionable.py (new): Actionable +
  ActionablePackageAvailableVersion +
  Repository_ActionablePackageAvailableVersion (Python identifier
  unchanged in this commit — Sprint 46.4 renames it).
- libinv/models/repository.py (new): Repository + Account +
  DeploymentCheckpoint (co-related, kept together).
- libinv/models/__init__.py: re-exports for everything above.
- libinv/models/_legacy.py: shrunk; ~17 orphaned imports dropped.

Test mock-patch redirects: 1 in test_wasp_eat_caterpillar.py, 5 in
test_actionable_populate_bulk.py, 1 in test_models_scio_http_path.py.
The audit was systematic — every `mock.patch("libinv.models._legacy.<
extracted_class>")` redirected to `libinv.models.<new_module>.<class>`.

Sprint 41.4 (Secbug + Sast) + Sprint 41.5 (verify __init__.py < 50
lines, mostly re-exports) close out Phase 3.

Verification: 94 passed (was 90 — Sprint 46.3 EPSS pruning landed in
parallel and adds 4 new tests), 3 skipped, 1 xfailed. mypy clean
across libinv/models/ (10 source files now). Import smoke test passes.
